widen findPlayerItem itemId to uint32_t and scan all inventory slots
matches Item::getId()/findItemInContainers (uint32_t) to avoid truncating
32-bit item ids, and scans up to LastInventorySlot so the helper is correct
for callers on modern protocols too. addresses review feedback on #1541.

+++ src/client/game.h
@@ -186,7 +186,7 @@ public:
     void useInventoryItem(uint16_t itemId);
     void useInventoryItemWith(uint16_t itemId, const ThingPtr& toThing);
     ItemPtr findItemInContainers(uint32_t itemId, int subType, uint8_t tier);
-    ItemPtr findPlayerItem(uint16_t itemId, int subType);
+    ItemPtr findPlayerItem(uint32_t itemId, int subType);
 
     // container related
     int open(const ItemPtr& item, const ContainerPtr& previousContainer);
